+++ app/services/insights/average_time/for_review_service.rb
@@ -38,6 +38,7 @@ module Insights
         return 1 if created_at >= reviewed_at
         return reviewed_at.to_i - created_at.to_i unless @insightable.with_work_time?
 
+        find_using_work_time(pull_request)
         seconds_between_times(
           convert_time(created_at),
           convert_time(reviewed_at),